+++ NEWS.md
@@ -7,6 +7,7 @@
 * Closed #227: bs_themer() now overlays a spinner during Sass compilation (#243).
 * Closed #289: collapsed navbar toggle now correctly floats to the right (#290) 
 * Closed rstudio/flexdashboard#316: fixed an issue with navbar nav spacing/alignment (#286).
+* Closed #279: Warnings about low color contrasts are now suppressed by default, unless `shiny::devmode()` is enabled. To enable/disable these warnings, set the new `options(bslib.color_contrast_warnings = )` to `TRUE`/`FALSE` (#287).
 
 # bslib 0.2.4